+++ src/test/resources/sectionDao.sql
@@ -6,7 +6,7 @@ create table SECTION
     line_id         bigint not null,
     up_station_id   bigint not null,
     down_station_id bigint not null,
-    distance        bigint not null,
+    distance        int    not null,
     primary key (id)
 --     foreign key (line_id) references LINE(id),
 --     foreign key (up_station_id) references  STATION(id),
